refactor: modularize ditbinmas report generation
